refactor: deduplicate Parser struct into shared sys/parse module
Move the identical Parser struct from mio.rs and tty.rs into
sys/parse.rs. Add push_event() method for Windows hybrid source to
enqueue non-ANSI events alongside parsed events.
